refactor: Move `converter_helpers.py` functions into `converter.py`
- `converter.py` only had one external facing function that called an internal function with the same parameters.
- Not sure if there was a specific reason for this setup, can be undone if needed.
